+++ databricks/sdk/errors/__init__.py
@@ -1,5 +1,4 @@
 from .base import DatabricksError, ErrorDetail
-from .mapper import _error_mapper
 from .parser import _Parser
 from .customizer import _ErrorCustomizer
 from .platform import *
